Added saving to file and other small refactors
Added saving sorted output into file (-o file).
More work to allow sorting by count. Right know
some plumbing is there but we need to actually
use it in print_map_loop(). For now we use todo!()
to panic in the print thread - but we still do full
counting and we save into a file if one is passed.

Signed-off-by: Marek Kroemeke <kroemeke@gmail.com>